Update README.md
Great heroes and diabolical villains are the types of characters that people long to be—but not Cid Kagenou. He longs to become the true mastermind behind it all, pulling the strings, and having his machinations drive the entire story. In his previous life in modern-day Japan, he was not able to achieve his goals. But now, reborn into a world of magic, he will show the true eminence of shadow! As Cid pretends to be just another average mob character, he recruits members to his (fake) organization Shadow Garden, aiming to take down the mysterious cult who lurks in the shadows. With plenty of ad-libs, Cid leads his (fake) organization against the foul (yet surprisingly real) Diabolos Cult.